Organize src/lib/game/ directory into domain subdirectories #26
