+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@joshuafolkken/game-kit",
-	"version": "0.12.0",
+	"version": "0.13.0",
 	"keywords": [
 		"svelte"
 	],
