+++ src/lib/game/GameScene.svelte.spec.ts
@@ -1,9 +1,9 @@
 import { audio } from '$lib/game/audio'
 import { device } from '$lib/game/device.svelte'
-import { fullscreen_switch_input } from '$lib/game/fullscreen-switch-input'
 import { fullscreen } from '$lib/game/fullscreen.svelte'
 import { session } from '$lib/game/session.svelte'
 import { game_state } from '$lib/game/state.svelte'
+import { fullscreen_switch_input } from '$lib/game/switch/fullscreen-switch-input'
 import { flushSync } from 'svelte'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { render } from 'vitest-browser-svelte'
