+++ src/lib/game/controls/ControlsOverlay.svelte
