+++ src/lib/game/controls/ControlsOverlay.svelte.spec.ts
