+++ src/lib/game/controls/JumpIcon.svelte
