+++ src/lib/game/controls/JumpIcon.svelte.spec.ts
