+++ src/lib/game/controls/KeyboardDiagram.svelte
