+++ src/lib/game/controls/KeyboardDiagram.svelte.spec.ts
