+++ src/lib/game/controls/MouseDiagram.svelte
