+++ src/lib/game/controls/MouseDiagram.svelte.spec.ts
