+++ src/lib/game/controls/TouchDiagram.svelte
