+++ src/lib/game/controls/TouchDiagram.svelte.spec.ts
