+++ src/lib/game/controls/VirtualJoystick.svelte.spec.ts
