+++ src/lib/game/display/fps.svelte.spec.ts
