+++ src/lib/game/display/fps.svelte.ts
