+++ src/lib/game/display/score-display-config.spec.ts
