+++ src/lib/game/display/score-display-config.ts
