+++ src/lib/game/display/score-display-types.ts
