+++ src/lib/game/scene/FloorCredits.svelte.spec.ts
