+++ src/lib/game/scene/Room.svelte
