+++ src/lib/game/scene/Room.svelte.spec.ts
