+++ src/lib/game/scene/credits-config.spec.ts
