+++ src/lib/game/scene/credits-config.ts
