+++ src/lib/game/scene/lighting.spec.ts
