+++ src/lib/game/scene/lighting.ts
