+++ src/lib/game/scene/room-config.ts
