+++ src/lib/game/scene/scene-objects-config.ts
