+++ src/lib/game/switch/alt-switch-input.ts
@@ -1,5 +1,5 @@
 import { game_state } from '$lib/game/state.svelte'
-import { create_switch_input } from '$lib/game/switch-input'
+import { create_switch_input } from '$lib/game/switch/switch-input'
 
 const { on_click } = create_switch_input({ action: () => game_state.toggle_alt() })
