+++ src/lib/game/switch/fps-switch-input.ts
@@ -1,5 +1,5 @@
-import { fps } from '$lib/game/fps.svelte'
-import { create_switch_input } from '$lib/game/switch-input'
+import { fps } from '$lib/game/display/fps.svelte'
+import { create_switch_input } from '$lib/game/switch/switch-input'
 
 const { on_click } = create_switch_input({ action: () => fps.toggle() })
