+++ src/lib/game/switch/switch-colors.spec.ts
