+++ src/lib/game/switch/switch-colors.ts
