+++ src/lib/game/switch/switch-config.spec.ts
