+++ src/lib/game/switch/switch-config.ts
