+++ src/routes/+layout.svelte
@@ -3,7 +3,7 @@
 	import click_url from '$lib/assets/sound/dragon-studio-distorted-electronic-click-472367.opus'
 	import { loading, OVERLAY_ELEMENT_ID, OVERLAY_HIDDEN_CLASS } from '$lib/game/loading.svelte'
 	import { game_state } from '$lib/game/state.svelte'
-	import { switch_audio } from '$lib/game/switch-audio'
+	import { switch_audio } from '$lib/game/switch/switch-audio'
 	import { messages } from '$lib/messages/en'
 
 	const LOADING_STATUS_ID = 'loading-status'
